requirements.txt
fastapi
sqlalchemy
datetime
jose
azure-storage-blob
